updated: restructured the project layout. added: documentation

+++ db/deletetaskifowned.go
@@ -6,23 +6,36 @@ import (
 	"fmt"
 )
 
+// DeleteTaskIfOwned deletes the task identified by taskID if it is owned by userID.
+//
+// If Repo is not initialized, returns an ErrNotConnected.
+// If taskID is empty or no task matches, returns an ErrTaskNotExist.
+// If userID is empty, returns an ErrUserNotExist.
+// If an error occurs in the Repo, returns an ErrInternalRepo.
+// If the task is not owned by userID, returns an ErrNotOwner.
+// If no task was completed, but the task exists and is owned by userID, returns an ErrorUnknown. This should never happen.
 func (r *Repo) DeleteTaskIfOwned(taskID, userID int) (err error) {
 	var (
-		caller       = "DeleteTaskIfOwned"
 		res          sql.Result
 		rowsAffected int64
 		t            Task
 	)
 	if !r.isConnected() {
-		return fmt.Errorf("%s: %w", caller, ErrNotConnected)
+		return ErrNotConnected
+	}
+	if taskID < 1 {
+		return fmt.Errorf("%w for id %d", ErrTaskNotExist, taskID)
+	}
+	if userID < 1 {
+		return fmt.Errorf("%w for id %d", ErrUserNotExist, userID)
 	}
 
 	res, err = r.db.Exec(`DELETE FROM task WHERE id = ? AND user_id = ?`, taskID, userID)
 	if err != nil {
-		return fmt.Errorf("%s: %w", caller, NewErrRepo(err))
+		return fmt.Errorf("%w: %s", ErrInternalRepo, err)
 	}
 	if rowsAffected, err = res.RowsAffected(); err != nil {
-		return fmt.Errorf("%s: %w", caller, NewErrRepo(err))
+		return fmt.Errorf("%w: %s", ErrInternalRepo, err)
 	}
 
 	// Happy path
@@ -32,14 +45,14 @@ func (r *Repo) DeleteTaskIfOwned(taskID, userID int) (err error) {
 
 	if t, err = r.GetTaskByID(taskID); err != nil {
 		if errors.Is(err, sql.ErrNoRows) {
-			return fmt.Errorf("%s: %w", caller, NewErrTaskNotExist(taskID))
+			return fmt.Errorf("%w for id %d", ErrTaskNotExist, taskID)
 		}
-		return fmt.Errorf("%s: %w", caller, err)
+		return fmt.Errorf("%w: %s", ErrInternalRepo, err)
 	}
 
 	if t.UserID != userID {
-		return fmt.Errorf("%s: %w", caller, NewErrNotOwner(userID, taskID))
+		return ErrNotOwner
 	}
 
-	return fmt.Errorf("%s: %w", caller, ErrUnknown)
+	return ErrUnknown
 }
